move configuration.ts to services package
