+++ src/functions/authoriser.ts
@@ -3,7 +3,7 @@ import StatementBuilder from "../services/StatementBuilder";
 import {JWTService} from "../services/JWTService";
 import {StatusCodeError} from "request-promise/errors";
 import IConfig from "../utils/IConfig";
-import configuration from "../utils/configuration";
+import configuration from "../services/configuration";
 import {APIGatewayAuthorizerResult} from "aws-lambda/trigger/api-gateway-authorizer";
 
 /**
